# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Multi-Agent AI System",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/agent/{agent_id}/capability-issues": {
      "get": {
        "tags": [
          "agent"
        ],
        "summary": "Get Capability Issues",
        "description": "Get capability issues for an agent - PROTECTED (requires ownership)",
        "operationId": "get_capability_issues_api_v1_agent__agent_id__capability_issues_get",
        "parameters": [
          {
            "name": "agent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Agent Id"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by status: open or resolved",
              "title": "Status"
            },
            "description": "Filter by status: open or resolved"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "description": "Max issues to return",
              "default": 100,
              "title": "Limit"
            },
            "description": "Max issues to return"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Number of issues to skip",
              "default": 0,
              "title": "Offset"
            },
            "description": "Number of issues to skip"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/agent/{agent_id}/capability-issues/resolve-all": {
      "post": {
        "tags": [
          "agent"
        ],
        "summary": "Resolve All Capability Issues",
        "description": "Bulk resolve all open capability issues for an agent - PROTECTED",
        "operationId": "resolve_all_capability_issues_api_v1_agent__agent_id__capability_issues_resolve_all_post",
        "parameters": [
          {
            "name": "agent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Agent Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/agent/capability-issues/{issue_id}/resolve": {
      "post": {
        "tags": [
          "agent"
        ],
        "summary": "Resolve Capability Issue",
        "description": "Resolve a single capability issue - PROTECTED (requires ownership)",
        "operationId": "resolve_capability_issue_api_v1_agent_capability_issues__issue_id__resolve_post",
        "parameters": [
          {
            "name": "issue_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Issue Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/agent/getAgent/{agent_id}": {
      "get": {
        "tags": [
          "agent"
        ],
        "summary": "Get Agent",
        "description": "Get agent by ID - PUBLIC (authentication optional)",
        "operationId": "get_agent_api_v1_agent_getAgent__agent_id__get",
        "parameters": [
          {
            "name": "agent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Agent Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/agent/getAllAgents": {
      "get": {
        "tags": [
          "agent"
        ],
        "summary": "Get Agent List",
        "description": "Get visible agents, optionally filtering to active agents.",
        "operationId": "get_agent_list_api_v1_agent_getAllAgents_get",
        "parameters": [
          {
            "name": "active_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Return only active agents when true",
              "default": false,
              "title": "Active Only"
            },
            "description": "Return only active agents when true"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}